modified:   a11y-configs/pi-arch/console-sd-libao/scripts/120-build-espeak.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/999-finalize.sh

+++ a11y-configs/pi-arch/console-sd-libao/scripts/999-finalize.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 echo '-- Finished everything, removing the build path directory...'
-rm -rf "${BUILD_PATH}"
+[ "${SM_TIDY}" ] && rm -rf "${BUILD_PATH}"
 aplay /usr/share/sounds/sound-icons/trumpet-12.wav &>/dev/null
 exit 0
